refactor: getRankStatistics 메서드를 lottoGame.js로 이동
