+++ src/step1-index.js
@@ -1,6 +1,6 @@
 import { readLineAsync } from './utils/readlineAsync.js'
 import { getTicket, LOTTO_PRICE } from './lotto.js'
-import { calculateProfitRate, getRank, LOTTO_RANK_TABLE } from './lottoGame.js'
+import { calculateProfitRate, getRank, getRankStatistics } from './lottoGame.js'
 import { printProfit, printResults, printTicket } from './lottoPrint.js'
 
 async function start() {
@@ -31,29 +31,7 @@ async function getWinningNumber() {
 
 function getStatistics(tickets, winningNumbers, winningBonusNumber) {
   const results = tickets.map(ticket => getRank(ticket, winningNumbers, winningBonusNumber)).filter(Boolean)
-
   return getRankStatistics(results)
-}
-
-function getRankStatistics(results) {
-  let totalWinnings = 0
-  const stats = LOTTO_RANK_TABLE.map(rank => ({
-    ...rank,
-    count: 0,
-  }))
-
-  results.forEach(result => {
-    totalWinnings += result.prize
-    const rankIndex = stats.findIndex(rank =>
-      rank.matchCount === result.matchCount && rank.bonus === result.bonus
-    )
-
-    if (rankIndex !== -1) {
-      stats[rankIndex].count += 1
-    }
-  })
-
-  return { stats: stats.reverse(), totalWinnings }
 }
 
 await start()
